Write restart at end of run via NUOPC component attribute & Initialize cpl_scalar field when created for UFS (#969)
These are two commits cherry-picked from as in UFS and needed to close NOAA-EMC#84.

This PR adds the ability for CICE to write restart files at the end of the run (independent of other settings) and controlled via the CMEPS configuration option write_restart_at_endofrun. Setting this configuration option to True creates a restart file in the same way for CMEPS, MOM6, and CICE.

This PR also initializes the scalar field value for the index for NTile (implemented for FV3) not used or set in CICE. In certain cases, the scalar field value for this index has been found to be non-zero (NaN in debug compiles). This is the cause of the failure reported in ufs-community/ufs-weather-model#2338.

* Add end of run functionality to CICE (#77)
* initialize cpl_scalar field when created (#83)

---------

Co-authored-by: Daniel Sarmiento <42810219+dpsarmie@users.noreply.github.com>
Co-authored-by: Denise Worthen <denise.worthen@noaa.gov>